+++ flink-runtime/src/test/java/org/apache/flink/runtime/testutils/CommonTestUtils.java
@@ -45,82 +45,6 @@ import org.apache.flink.core.memory.OutputViewDataOutputStreamWrapper;
 public class CommonTestUtils {
 
 	/**
-	 * Creates a copy of the given {@link IOReadableWritable} object by an in-memory serialization and subsequent
-	 * deserialization.
-	 * 
-	 * @param original
-	 *        the original object to be copied
-	 * @return the copy of original object created by the original object's serialization/deserialization methods
-	 * @throws IOException
-	 *         thrown if an error occurs while creating the copy of the object
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T extends IOReadableWritable> T createCopyWritable(final T original) throws IOException {
-
-		final ByteArrayOutputStream baos = new ByteArrayOutputStream();
-		final DataOutputStream dos = new DataOutputStream(baos);
-
-		original.write(new OutputViewDataOutputStreamWrapper(dos));
-
-		final String className = original.getClass().getName();
-
-		Class<T> clazz = null;
-
-		try {
-			clazz = (Class<T>) Class.forName(className);
-		} catch (ClassNotFoundException e) {
-			fail(e.getMessage());
-		}
-
-		T copy = null;
-		try {
-			copy = clazz.newInstance();
-		} catch (Throwable t) {
-			t.printStackTrace();
-			fail(t.getMessage());
-		}
-
-		final ByteArrayInputStream bais = new ByteArrayInputStream(baos.toByteArray());
-		final DataInputStream dis = new DataInputStream(bais);
-
-		copy.read(new InputViewDataInputStreamWrapper(dis));
-		if (dis.available() > 0) {
-			throw new IOException("The coped result was not fully consumed.");
-		}
-
-		return copy;
-	}
-	
-	@SuppressWarnings("unchecked")
-	public static <T extends java.io.Serializable> T createCopySerializable(T original) throws IOException {
-		if (original == null) {
-			throw new IllegalArgumentException();
-		}
-		
-		ByteArrayOutputStream baos = new ByteArrayOutputStream();
-		ObjectOutputStream oos = new ObjectOutputStream(baos);
-		oos.writeObject(original);
-		oos.close();
-		baos.close();
-		
-		ByteArrayInputStream bais = new ByteArrayInputStream(baos.toByteArray());
-		ObjectInputStream ois = new ObjectInputStream(bais);
-		
-		T copy;
-		try {
-			copy = (T) ois.readObject();
-		}
-		catch (ClassNotFoundException e) {
-			throw new IOException(e);
-		}
-		
-		ois.close();
-		bais.close();
-		
-		return copy;
-	}
-
-	/**
 	 * Sleeps for a given set of milliseconds, uninterruptibly. If interrupt is called,
 	 * the sleep will continue nonetheless.
 	 *
